requirements.txt
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-language==2.14.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
zipp==3.20.0
